+++ cratedb_toolkit/sqlalchemy/__init__.py
@@ -1,2 +1,1 @@
-from .patch import patch_inspector
 from .polyfill import check_uniqueness_factory, polyfill_autoincrement, polyfill_refresh_after_dml, refresh_table